Match LimitedAvailability to new Unauthenticated
